refactor: change type file
