+++ apps/web/src/common/modules/widgets/_widget-fields/WidgetFieldCategoryBy.vue
@@ -7,8 +7,9 @@ import {
 import type { MenuItem } from '@spaceone/design-system/types/inputs/context-menu/type';
 
 import { useProxyValue } from '@/common/composables/proxy-state';
-import type { CategoryByOptions } from '@/common/modules/widgets/types/widget-config-type';
-import type { WidgetFieldComponentProps, WidgetFieldComponentEmit } from '@/common/modules/widgets/types/widget-field-type';
+import type {
+    WidgetFieldComponentProps, WidgetFieldComponentEmit, CategoryByOptions,
+} from '@/common/modules/widgets/types/widget-field-type';
 
 
 const props = withDefaults(defineProps<WidgetFieldComponentProps<CategoryByOptions>>(), {
