+++ apps/web/src/common/modules/widgets/_widget-fields/WidgetFieldTotalField.vue
@@ -5,8 +5,11 @@ import { PSelectDropdown, PFieldGroup } from '@spaceone/design-system';
 import type { MenuItem } from '@spaceone/design-system/types/inputs/context-menu/type';
 
 import { useProxyValue } from '@/common/composables/proxy-state';
-import type { TotalFieldOptions } from '@/common/modules/widgets/types/widget-config-type';
-import type { WidgetFieldComponentProps, WidgetFieldComponentEmit } from '@/common/modules/widgets/types/widget-field-type';
+import type {
+    WidgetFieldComponentProps,
+    WidgetFieldComponentEmit,
+    TotalFieldOptions,
+} from '@/common/modules/widgets/types/widget-field-type';
 
 
 const props = withDefaults(defineProps<WidgetFieldComponentProps<TotalFieldOptions>>(), {
